Webcord: fix up and reimplement installer

+++ scripts/discord.sh
@@ -8,41 +8,45 @@ cd /tmp
 rm webcord*
 
 if command -v apt >/dev/null; then
-  echo "Installing dependencies..."
-  sudo apt install git curl -y || error "Couldn't install dependencies"
+  if [[ $dpkg_architecture == "amd64" ]] || [[ $dpkg_architecture == "arm64" ]] || [[ $dpkg_architecture == "armhf" ]]; then
+    echo "Installing dependencies..."
+    python3 -m pip --install --upgrade pip lastversion || error "Couldn't install dependencies"
+    sudo apt install git curl -y || error "Couldn't install dependencies"
 
-  echo "Removing previous legacy Discord installs and inconsistent apt repo..."
-  sudo dpkg -r electron-discord-webapp
-  sudo rm /etc/apt/sources.list.d/webcord.list*
+    echo "Removing previous legacy Discord installs and inconsistent apt repo..."
+    sudo dpkg -r electron-discord-webapp
+    sudo rm /etc/apt/sources.list.d/webcord.list*
 
-  echo "Downloading the most recent .deb from SpacingBat3's repository..."
-  #note to self: fix this to be cross-architecture later
-  curl -s https://api.github.com/repos/SpacingBat3/WebCord/releases/latest |
-    grep "browser_download_url.*arm64.deb" |
-    cut -d : -f 2,3 |
-    tr -d \" |
-    wget -i -
+    echo "Downloading the most recent .deb from SpacingBat3's repository..."
+    lastversion --assets --filter $dpkg_architecture.deb download https://github.com/SpacingBat3/WebCord
 
-  echo "Done! Installing the package..."
-  sudo apt install -y /tmp/*arm64.deb || error "Webcord install failed"
+    echo "Done! Installing the package..."
+    sudo apt install -y /tmp/webcord*$dpkg_architecture.deb || error "Webcord install failed"
+  else
+    error_user "Your architecture ($dpkg_architecture) is not supported."
+  fi
 
-elif command -v dnf >/dev/null; then
-  echo "Installing dependencies..."
-  sudo dnf install git curl -y || error "Couldn't install dependencies"
+elif
+  command -v dnf >/dev/null
+then
+  # I'm not including 32-bit arm here despite Webcord (currently) providing downloads for it because
+  # 1. Fedora deprecated it https://arm.fedoraproject.org/
+  # 2. I'm too lazy to parse out armv7l and armv7hl
+  if [[ $architecture == "arm64" ]] || [[ $architecture == "x86_64" ]]; then
+    echo "Installing dependencies..."
+    sudo yum -y install https://extras.getpagespeed.com/release-latest.rpm
+    sudo yum -y install lastversion
 
-  echo "Downloading the most recent .rpm from SpacingBat3's repository..."
-  #note to self: fix this to be cross-architecture later
-  curl -s https://api.github.com/repos/SpacingBat3/WebCord/releases/latest |
-    grep "browser_download_url.*arm64.rpm" |
-    cut -d : -f 2,3 |
-    tr -d \" |
-    wget -i -
+    echo "Downloading and installing the most recent .rpm from SpacingBat3's repository..."
+    # this command unfortunately doesn't work on Debian-based systems
+    sudo lastversion install webcord || error "Webcord install failed"
 
-  echo "Done! Installing the package..."
-  sudo dnf install -y /tmp/*arm64.rpm || error "Webcord install failed"
+  else
+    error_user "Your architecture ($architecture) is not supported."
+  fi
 
 else
-  error "No available package manager found. Are you using a Ubuntu/Debian or Fedora based system?"
+  error_user "No available package manager found. Are you using a Ubuntu/Debian or Fedora based system?"
 fi
 
 cd ~
